+++ CHANGELOG.md
@@ -34,6 +34,7 @@
 * Incoming transfer frames received during initial link detach are no longer discarded.
 * Session will no longer flood peer with flow frames when half its incoming window is consumed.
 * Newly created `Session` won't leak if the context passed to `Client.NewSession()` expires before exit.
+* Newly created `link` won't leak if the context passed to `link.attach()` expires before exit.
 
 ### Other Changes
 * Errors when reading/writing to the underlying `net.Conn` are now wrapped in a `ConnectionError` type.
@@ -42,3 +43,4 @@
 * Only send one flow frame when a drain has been requested.
 * Session window size increased to 5000.
 * Creation and deletion of `Session` instances have been made deterministic.
+* Allocation and deallocation of link handles has been made deterministic.